requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlmodel==0.0.8
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
redis==5.0.1
Pillow==10.1.0
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
python-multipart==0.0.6
websockets==12.0
pywebpush==1.14.1
pyjwt==2.8.0
